# Client SDK/go-client-generated/api/swagger.yaml
openapi: 3.0.2
info:
  title: OCS Inventory Rest API
  version: "1.0"
servers:
- url: https://api.ocsinventory-ng.tld/ocsapi/v1
security:
- basicAuth: []
paths:
...
  /snmp/{snmp_type}:
    get:
      summary: Get snmp device list of a specific snmp type
      parameters:
      - name: snmp_type
        in: path
        description: Name of the snmp type you want to list
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: start
        in: query
        description: Start offset of the query
        required: false
        style: form
        explode: true
        schema:
          type: integer
      - name: limit
        in: query
        description: Limit offset of the query
        required: false
        style: form
        explode: true
        schema:
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SNMPDetail'
        "401":
          description: Authentication information is missing or invalid
          headers:
            WWW_Authenticate:
              style: simple
              explode: false
              schema:
                type: string
                description: Authorization token configured in Apache2 host
      security:
      - basicAuth: []
  /snmp/{snmp_type}/{id}:
    get:
      summary: Get a device from a snmp type with its ID
      parameters:
      - name: snmp_type
        in: path
        description: Name of the snmp type you want to list
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: id
        in: path
        description: Identifiant en base de données de l'équipement
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SNMPDetail'
        "401":
          description: Authentication information is missing or invalid
          headers:
            WWW_Authenticate:
              style: simple
              explode: false
              schema:
                type: string
                description: Authorization token configured in Apache2 host
      security:
      - basicAuth: []
...
components:
  schemas:
...
    SNMPDetail:
      type: array
      items:
        $ref: '#/components/schemas/SNMPDetail_inner'
      x-schema-name: SNMPDetail
...
    SNMPDetail_inner:
      type: object
      properties:
        DefaultAddressIP:
          type: string
          example: 127.0.0.1
        DefaultDescription:
          type: string
          example: This is a description
        DefaultGateway:
          type: string
          example: 172.18.25.254
        DefaultLocation:
          type: string
          example: Here
        DefaultMacAddress:
          type: string
          format: nullable
        DefaultName:
          type: string
          example: My Equipment
        DefaultUptime:
          type: string
          format: nullable
        ID:
          type: integer
          format: int32
          example: 1
        LASTDATE:
          type: string
          format: date-time
          example: 2023-06-08T12:22:18Z
      example:
        DefaultAddressIP: 127.0.0.1
        DefaultDescription: This is a description
        LASTDATE: 2023-06-08T12:22:18Z
        DefaultUptime: DefaultUptime
        DefaultGateway: 172.18.25.254
        DefaultName: My Equipment
        DefaultMacAddress: DefaultMacAddress
        ID: 1
        DefaultLocation: Here
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic